refactor: split app mode key handlers
